refactor(app): moved `callback.types.ts` to `types` dir from `hooks` dir
