+++ app/src/hooks/callbacks/useJoinGameCallback.ts
@@ -3,7 +3,7 @@ import { useZkubeContract } from '../useContract';
 import { usePrivyWalletAddress } from '../usePrivyWalletAddress';
 import { useChainId, useChains } from 'wagmi';
 import { StringNumberBI } from '../../types/Puzzle';
-import { ContractCallbackReturnType } from './types';
+import { ContractCallbackReturnType } from '../../types/callbacks.types';
 
 export function useJoinGameCallback() {
   const zKubeContract = useZkubeContract(true);
